requirements.txt
# Developing
Werkzeug
django-debug-toolbar
django-extensions

# Testing
mock>=1.0.1,<1.1
django-nose>=1.1,<1.2
pinocchio>=0.3.1,<=0.4
coverage>=3.6,<3.7
